+++ inst/examples/devel.cvmeasure.R
@@ -21,8 +21,6 @@ cmp <- coordinates ~ vegetation(map = gorillas$gcov$vegetation, model = "factor"
   spde(map = coordinates, model = pcmatern, mesh = gorillas$mesh) -
   Intercept
 
-init.tutorial() # Faster inferece for example
-
 fit <- lgcp(cmp, gorillas$nests, samplers = gorillas$boundary)
 
 # Predict SPDE and vegetation at the mesh vertex locations